Assign positive and negative emojis and add initial code for checking the position of emojis and iphone
