+++ play.html
@@ -48,5 +48,6 @@
 <script src="https://cdnjs.cloudflare.com/ajax/libs/twitter-bootstrap/4.6.0/js/bootstrap.min.js"></script>
 <script src="assets/js/phone-slider.js"></script>
 <script src="assets/js/emoji_drop.js"></script>
+<script src="assets/js/scoring.js"></script>
 </body>
 </html>
